Added wavefront using pull terminals as a separate example
